requirements.txt
annotated-types==0.7.0
anyio==4.7.0
certifi==2024.12.14
cffi==1.17.1
charset-normalizer==3.4.1
click==8.1.8
colorama==0.4.6
cryptography==44.0.0
deepdiff==8.1.1
distro==1.9.0
et_xmlfile==2.0.0
fastapi==0.115.6
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
idna==3.10
jiter==0.8.2
lxml==5.3.0
numpy==2.2.1
openai==1.58.1
openpyxl==3.1.5
orderly-set==5.2.3
pandas==2.2.3
pdfminer.six==20231228
pdfplumber==0.11.4
pillow==11.0.0
pycparser==2.22
pydantic==2.10.4
pydantic_core==2.27.2
PyMuPDF==1.25.1
PyPDF2==3.0.1
pypdfium2==4.30.1
python-dateutil==2.9.0.post0
python-docx==1.1.2
python-dotenv==1.0.1
python-multipart==0.0.20
pytz==2024.2
setuptools==75.6.0
six==1.17.0
sniffio==1.3.1
starlette==0.41.3
tqdm==4.67.1
typing_extensions==4.12.2
tzdata==2024.2
uvicorn==0.34.0
psycopg2==2.9.10
sqlalchemy==2.0.36
pydantic-settings==2.4.0
logging==0.4.9.6
datetime==5.5
typing
passlib==1.7.4
python-jose==3.3.0
alembic==1.13.2
psycopg2-binary==2.9.10
